New approach for generalizing RT FEs for arbitrarily oriented meshes (documentation added to RaviartThomasRefFEs.jl)
Two key ingredients in the implementation of this type of ReferenceFE are the
get_shapefuns(model,cell_reffes) and get_dof_basis(mode,cell_reffes) overloads.
These are written such that, for each cell K, they return the shape functions
and dof values in the *global* RT space. For a dof owned by a face which is shared by
two cells, there is a master and a slave cell. The slave cell first computes the
shape functions and dof values using local-to-cell data structures, but then flips the
sign of both in order to get their corresponding counterparts in the **global**
RT space. As as result we have the following:

* When we interpolate a function into the global FE space, and we perform the cell-wise
  DoF values to global DoF values gather operation, we can either extract the global DoF value from
  the master or slave cell without worrying about the sign.
* When we evaluate a global FE function, and we perform the global DoF values to
  cell-wise DoF values scatter operation, we don't have to worry about the sign either.
  On the slave cell, we will have both the sign of the DoF value, and the sign of the
  shape function corresponding to the global DoF.
* We do NOT have to use the signed determinant, but its absolute value, in the Piola Map.