Completely refactored LinearizedFESpace towards higher efficiency.
As opposed to the (highly ineffecient) previous implementation,
now LinearizedFESpace is defined on the refined triangulation
(and no longer on the coarse triangulation).

This decision produces some violations of the invariants inherent
to collect_* functions that I had to bypass by overloading the
get_fe_cell_data functions.

As positive notes:
  1. the get_gradient_fe_space() function is no longer required.
  2. We do not longer need CompositeMeasure.