RefinementRule typing now homogeneous for mixed.
In order to work with lazy arrays, we need
cell-arrays of RefinementRules to have a
concrete and homogeneous typing.

For meshes with mixed refining, like
edge-based refined meshes, that was not
the case. This is now fixed.

Now arrays of FineToCoarseFields and
FineToCoarseReferenceFEs have concrete
eltype as well, despite containing a
RefinementRule.