[wip] added indep_comp_getindex and fix FESpaces
of value with linked components.
Indeed, the code previously made the assumption that all the components
of the Number unknown of the FESpaces are unlinked to the other, which
is false for the symmetric tensor types.

A new getter for the independent components of the Number is added and
used in the MultiValue'd FESpaces machinery. Also, the number of DoFs of the
ReferenceFEs are given by num_indep_components(::Number).

fixes #923 #908