Performance optimization. The test on the injectivity of the tface_to_mface_map of BodyFittedTriangulation is moved to the BodyFittedTriangulation constructor, so that it is performed only once and not each time we call get_glue. Added a trait to BodyFittedTriangulation that stores the result of the test on injectivity.
Besides, for correctness, the mface_to_tface_map is set to nothing
as currently PosNegPartition is not prepared to handle non-injective
mappings. This will have to be replaced by a proper data structure
that is able to handle non-injective mappings if in the future
we find an scenario in which this inverse mapping is required to
be used.