optimization of trivial MulAddMap evaluate!
The (0,1) AttachDirichletMap is just (A,b,c) -> c .

So we can avoid `mul!` in this case (but keep the `copyto!` for type
stability when c is real but b and d=cache[2] are complex).

I believe this is safe regarding type stability since
`return_cache(::AttachDirichletMap)` uses an eager `Base.*` to create
the `cache`, it does not rely on this evaluate.