Bugfix in topology for non-convex polytopes
In the case of non-convex general polytopes, you can get
into the following situation: Given a cell, and a face on
that cell, all the vertices of the face also belong to
a neighboring cell but the face itself does not. This is
not possible with normal tetrahedra/hexahedra, but it is
possible with non-convex polytopes, where one polytope
can wrap around another one.